Avoid a warning.
git-svn-id: https://svn.dealii.org/trunk@20052 0785d39b-7218-0410-832d-ea1e28bc413d

+++ deal.II/deal.II/include/dofs/dof_accessor.templates.h
@@ -2355,7 +2355,9 @@ namespace internal
 		    typename BaseClass::ExcInvalidObject());
 	    Assert (&accessor.get_fe() != 0,
 		    typename BaseClass::ExcInvalidObject());
-	    Assert (local_source_end-local_source_begin == accessor.get_fe().dofs_per_cell,
+	    Assert (static_cast<unsigned int>(local_source_end-local_source_begin)
+		    ==
+		    accessor.get_fe().dofs_per_cell,
 		    typename BaseClass::ExcVectorDoesNotMatch());
 	    Assert (accessor.dof_handler->n_dofs() == global_destination.size(),
 		    typename BaseClass::ExcVectorDoesNotMatch());
